move get user profile out of the route handler
